Initial splinter node implementation
The node can be started by loading the requried certs and keys
for the certificate authority, server/clients sessions as well
as set up the ports the node will accept new connections on.
A list of peers will also be accepted the node will try to
connect to.

Node only the service port is currently used. Each new connection
will be handled by its own thread.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>