Implement CircuitCreateRequest handling and adding new connection
Handles recieving create circuit requests and adds a new connection
to the corresponding nodes if they do not exist.

SplinterDaemon should be responsible for creating and starting
new connections, so `Sender<(String, String)>` was added to
`Connection` that links back to the Daemon (parent thread). When
a connection is handling a request that would result in a new
connection being started it will send the IP to its parent
thread and that would call
`thread::spawn(move || connectin.handle_msg()`.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>