Panic with correct line number on poisoned lock
Calling .unwrap() or .expect() on a lock that has been poisoned will
cause the current thread to panic and a panic message to be printed. The
actual panic!() call is within .unwrap() or .expect(), so the line
number that gets printed as part of this panic is within the
implementation of .unwrap() or .expect(). This is not helpful when
debugging.

This PR adds two macros to libsplinter for "unwrapping" RwLock.read()
and RwLock.write() in a way that prints the line number where the
poisoned lock was encountered, which is much more helpful for debugging.
A macro is necessary to avoid the problem described above with .unwrap()
and .expect()

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>