Fix readiness behavior of registration
Previously, registering or reregistering set the readiness to write,
which may have accidentally removed a readiness state of read. If a
send occurred prior to a register, this caused an incorrect readiness
state.

This changes the behavior of register/reregister to retain the read
readiness state when setting the write state.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>