Implement SplinterState
SplinterState contains the storage location, the circuit state
that needs to be persisted, and the directory of service and the
node they are connected to.

When circuit state is updated, a new storage object is created
and the updated state object is written to the storage location.
This allows state to be cloned and passed around using an RwLock.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>