Hold on to peer write lock for the entirety of adding connection
Before it was possible for the connection to be added to mesh
before it was added to the peer list. This would cause network
to error on receiving the a message from this connection if it
was sent right after connection.

Also updates the error message for Send vs Recv error when
no peer is found.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>