Add docker and docker-compose files
Add a docker file, which runs the build in multiple stages, in order to
improve local build times, as well as final build results.  It first
sets ups the environment needed (rust nightly), and then creates an
empty project with a copy of the Cargo toml.  It runs the build, which
compiles are the dependencies.  Next, it removes the project
artifacts and copies in the source files.  This compiles the project
files only, which is a fast operations due to incremental compile.
Finally, it copies the built binary to a minimal image for small
publishing size.

Additionally, adds a docker compose file for convenience.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>