Implement NetworkMessageSender to send Messages from Handlers
A message handler will need a way to respond to messages.
To avoid having to clone the network into every handler, instead
they will be passed a sender end of a channel to send the
outgoing message to. The receiving end will be held by the
NetworkMessageSender. It will read off the outgoing messages
and send them over the network.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>