Add Dispatcher
This commit adds the dispatcher to the libsplinter network module. The
Dispatcher is configured with handlers by message type. These handlers
are executed when a message is dispatched.

The handler are implemented using a trait, which is generic on message
format. The Handler trait is implemented for the Fn trait with the
required parameters of the handle function.  This allows for handlers to
be implemented as either structs with the trait, named functions with
the correct parameters, and closures with the correct parameters.  All
three methods are demonstrated in the tests.

The message format must implement the trait `FromMessageBytes`. This
trait is implemented for protobuf::Message structs automatically, as
well as RawBytes (introduced due Rust not allowing trait specialization
yet).

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>