Add Message Context
This commit adds a message context struct that will be provided to all
handlers.  This struct contains information about the sent message: who
sent it, the type of message, and the raw bytes of the message,
initially.

This is useful for handlers that may want to reply to the message,
without needing any other way to get the peer information.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>