Move Receiver and Sender trait to channel module
Move out Receiver and Sender traits out of the network sender
and create mod called channel.

Also includes generic implementation for mpsc channels and
crossbeam channels that can be used in the network sender.
These implementations do not need to be publically available
and allow the client (for example splinterd) to use
crossbeam_channel directly.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>