Add CircuitDirectMessageHandler to handlers
If a CircuitDirectMessage is received, check that the circuit exists,
that the sender and recepient service are in the circuit roster and
have connected to the circuit. If so send the direct message either
to the connected service (recipient) or send to the node the service
is connected to.

Otherwise send a CircuitError message back to the sender of the
direct message.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>