Add CircuitErrorHandler to handlers
When a CircuitError message is not sent directly to a
service from the direct message handler, the node that
receives the message needs to try its best to send the
error message back to the correct service/node. If the
node is not able to figure out where to send it, the
message is dropped.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>