Add AuthorizationManager
Add the authorization manager which controls authorization states for a
connection.  These state transitions are a reflection of the types of
authorization actions that can occur.  Initially, these actions are
Connecting, and TrustIdentifying, and Unauthorizing.  Future actions can
be added and the state machine modified to take those into account.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>