Add NetworkAuthGuardHandler
Add a handler that wraps network message handlers with an authorization
manager.  This handler will only forward the messages to the wrapped
handler if the peer has completed the authorization message handshake.

The circuit messages and echo messages are wrapped in this guard when
splinterd's network message dispatcher is setup.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>