Add proto messages for n-phase commits
Add the proto messages for performing an n-phase commit with a generic
transaction.  The transaction payload is opaque to the request, and
verification is expected to be performed against a hash of the results
of the transaction's execution.

These messages make the assumption that the requesting node has produced
a correct answer and is looking for agreement from other members of the
circuit. As there is no followup message, currently, if the requestee
agrees, the expectation is that it has also applied the transaction.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>