Add InboundRouter and MessageFuture
The InboundRouter will take a correlation id and return a receiver
that can be used to block until a response is returned. The router
will send the response over the receiver if it has the correlation
id. If the router does not have a matching correlation id, the
message is sent over a default channel.

The MessageFuture wraps the Receiver and calling get will block
until the response is returned.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>