Add NPhaseCommit manager struct
The NPhaseCommit struct manages the status of a transaction through an
n-phase commit.  It records the expected verifying services, and their
current status (if pending or successful).

On a complete commit, it calls a given callback for the transaction.

If there is either a failure or a result mismatch, the transaction is
removed.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>