Create network and message handlers
Add functions which handle messages for

- Auth
- Circuit

The auth messages only handle a one-way handshake with the splinterd
service.  It does not require the splinter node authenticating with it
(as of this commit).

The circuit messages handled are registration and the specific messages
for the service.  These include the N-Phase message handling and
applying it to a stored state (this state exists in memory, as of this
commit).

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>