Add PeerMap for peer_id* -> mesh_id
Due to locking, a message may be sent via a channel for a temporary
peer, the peer id may be updated and the message maybe dropped due to
the temporary id no longer existing.

This commit introduces a PeerMap, which manages the relationship between
all peer ids and a mesh id over the lifetime of the connection. By
managing a forwarding table from a temp peer id to the actual peer_id,
and then the mesh id is determined from there.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>