Add peer_id to Services
The id is to be used as a peer_id when sending
a service a message from the node it is directly connected
too. Before that was done with the service id, but this
required that a service_id must be unique for all circuits.
With this change the same service_id can belong to multiple
circuits.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>