Reregister for writable when blocked
When writing to a transport is blocked, reregister for events for
when the transport will be writable.  This will ensure that messages
are sent when the connection is unblocked, even if there is something
not an inbound message on the same connection.

When the connection is writable again, reregister for readable only.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>