Add generating certs and whether to verify in tls to splinterd
When wanting to use tls in a development environment, the certs
can be automatically generated. Since the allowed ca certs will be
different everytime they are generated, the certs will not be
verified.

Also adds the ablity to turn tls verification off as a seperate
option and run in insecure mode.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>