Add initial grid-sdk with track and trace protos
This is the start of the grid-sdk that can be
used by developers to easily pull in other
grid components. Initially, it contains the track and
trace protos.

Also includes proto traits that will be used to
guard the rest of the code from using protobufs
directly in the future.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>