Rename PropertyValue to TrackAndTracePropertyValue
Schema and Track and Trace both have a protobuf message
called PropertyValue. This causes an issue when
generating the protobufs. Since Track and Trace
will be updated to use Grid Schema once it is
complete, this commit temporarily renames Track
and Trace's PropertyValue to TrackAndTracePropertyValue
so they can both build at the same time.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>