Create private XO network in docker compose
This commit updates the docker compose file for private XO to include
two splinterd containers and two private XO containers, with a circuit
definition shared between them.  The two private XO instances use each
other as the verifying services for their transactions.

This also includes a container that provides the xo client for
interacting with the various private services.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>