Add GridConfig and Builder
A GridConfig can be created via a Builder.  The intent of the GridConfig
builder is to join in values from various sources: default values and
cli args to start. Future commits may include joining values from a
configuration file.

This initial version includes values for the connect (i.e. validator
endpoint) and the logging level).

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>