Implement EventProcessor
The event processor spawns a thread which subscribes to and receives
Sawtooth event messages over a ZMQ connection.

As of this commit, the subscription is configured directly, specifically
tied to grid state addresses and block events.

The events themselves are handled by implementations of an EventHandler
trait.  These are passed in via the `start` method.

An event processor is shutdown and joined using the combination of an
EventProcessorShutdownHandle and JoinHandle.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>