Add REST API shutdown handle
This commit starts the rest api on its own thread, which allows other
threads and actions to be started in the main thread.  The run method
now returns a shutdown handle, used to signal a clean shutdown and a
join handle for the spawning thread to wait for the clean shutdown.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>