Add Pike to Grid docker-compose files
Updates the Pike Dockerfiles to be in line with the track and trace
Dockerfiles, and adds them to the Grid docker-compose and
docker-compose-installed files.

Removed a call Deb versioning script in the Pike Dockerfile-installed,
since that file does not exist in Grid. We may have to add a similar
script back in later.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>