Implement simple PermissionChecker
PermissionChecker is a helper that can be used to check if a Pike agent
has a certain permission. It can be imported from the Grid SDK into a
contract to execute Pike functionality.

This implementation is fairly simple, but in the future it can be
extended to allow smart permission checks.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>