Replace 2018 log use with macro_use crate stmt
Replace the 2018 style of use statements for the individual log macros
with the classic `macro_use` use-crate statement.  This is a minor
stylistic change, in that all use statements are similar (i.e. none of
them have a leading `::`).

In addition, the syntax causes issues in some editor in both formatting
rust code, as well as properly reading the file to determine symbol
names.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>