Check value was provided before creating new page
Currently, if a transaction is submitted in which all required
properties are in the payload, but some of the optional properties are
missing the transaction will fail, even though it should be valid. This
commit fixes that bug, by making sure that a property page is created
only when the property value is provided in the payload.

Signed-off-by: Eloá Franca Verona <eloafran@bitwise.io>