Add schema name to update grid property definition
To uniquely identify a grid property definition it is necessary the
property name and the schema name the property is part of.
The absence of schema_name in the
update_definition_end_block_num method could cause for
property definitions to be updated  when they are not supposed to be.

Signed-off-by: Eloá Franca Verona <eloafran@bitwise.io>