Update private counter to use two-phase consensus
Updates the private counter app to use the two-phase consensus
engine. This requires a number of changes:

- A new `PrivateCounterMessage` protobuf is introduced for communication
between private counter services; in addition to defining the protobuf
itself, the dependencies are updated, a `build.rs` is added, and the
Dockerfile is updated to build properly.
- The `PrivateCounterProposalManager` and `PrivateCounterNetworkSender`,
which implement the `ProposalManager` and `ConsensusNetworkSender`
traits respectively, are created for the service to handle interactions
with consensus.
- The service performs Message handling and increment creation using the
consensus mechanisms rather than handling consensus itself.

Signed-off-by: Logan Seeley <seeley@bitwise.io>