Add Sawtooth SDK dependency
This dependency is used for signing/keygen capabilities. The ZMQ library
is configured to reference a blank file since we don't need any of the
related functionality for sawtooth-sdk/signing, and it would require
additional configuration to use.

a d.ts. (TypeScript declaration) file is provided to enable use in
TypeScript modules.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>