Update private xo to use two-phase consensus
Updates the private xo app to use the two-phase consensus engine. This
requires a number of changes:

- A new `PrivateXoMessage` protobuf is introduced for communication
between private xo services; in addition to defining the protobuf
itself, the dependencies are updated, a `build.rs` is added, and the
Dockerfile is updated to build properly.
- The `PrivateXoProposalManager` and `PrivateXoNetworkSender`, which
implement the `ProposalManager` and `ConsensusNetworkSender` traits
respectively, are created for the service to handle interactions with
consensus.
- The service performs Message handling and proposal creation using the
consensus mechanisms rather than handling consensus itself.

Signed-off-by: Logan Seeley <seeley@bitwise.io>