Add implementation for network senders
This commit contains two implementations of network
senders that will be passed to the service after
the service completes connection with the splinter
daemon.

There is a StandardServiceNetworkSender that will be used
by most services and also an AdminServiceNetworkSender that
will be given to admin services running on the admin circuit.

Also includes a ServiceMessage enum that will be used to
send the circuit messages between threads in the
ServiceProcessor.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>