Add implementation for a network registry
This defines an implementation of a ServiceNetworkRegistry
that will be used by services to handle the messages required
to connect and disconnect a service with a splinter node. The
network registry will block until the response is received
from the node.

If the service connection is successful, a ServiceNetworkSender
is returned to the service and the service can start sending
messages at this time.

This includes hiding the circuit information from the service.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>