Add service processor implementation
The ServiceProcessor handles the networking for the
services that have been provided to the processor.
The processor handles connecting to the splinter node,
starting the service, and handling incoming and outgoing
messages from the splinter node.

The service will run in its own thread and be passed messages
by a channel.

Note this only works for inproc service currently. More
authorization message handling is required for the
ServiceProcessor to be used by an external service.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>