Add admin roster for admin circuits
Modify the Roster type to be an enum of either "Standard" or "Admin".
These two types of rosters have different ways of deciding if a services
is "contained" in the roster.  Any admin service that conforms to the
admin rules is accepted, where as Standard roster uses the previous
model of those set as a list of service names.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>