Send admin events to auth handlers by type
Send the AdminServiceEvents for proposals to the appropriate registered
auth handler, based on the circuit management type specified in the
proposal.

There can be any number of senders for a given type.  These are cleaned
up when an event fails to send, which would occur if the sender's
receiving end has dropped.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>