Implement deterministic coordinator for 2PC
Updates two phase commit to have an established coordinator; the
coordinator for a proposal is the verifier for that proposal with the
lowest peer ID. For standard proposals where all peers are verifiers,
the coordinator will be the same for all proposals.

This change is important because it prevents the network from wedging
itself when two nodes create proposals and request verification at the
same time; the network should only evaluate a single proposal at a time,
which is accomplished by having a single leader/coordinator.

When a node receives a proposal (whether created locally or received
from a peer), they check if they are the coordinator for that proposal
and if they are ready to start the consensus process. If not, they place
the proposal in the proposal backlog (and may attempt to start consensus
later on).

This commit also updates followers to only enter the EvaluatingProposal
state when a verification request is received.

Signed-off-by: Logan Seeley <seeley@bitwise.io>