Handle service networking in orchestrator
Rather than using service processors for running services, the
orchestrator handles networking on its own. The way the orchestrator
manages services is different from the service processor in the
following ways:
- The background threads used to send messages to and from services are
started when the orchestrator is created, not when services are started.
This is because an orchestrator is started at creation, and services are
initialized and shutdown on-the-fly.
- When a service is created, an background thread is not created to
receive messages from the inbound thread; instead, the inbound thread
passes the messages directly to the services. This allows the
orchestrator to maintain ownership of the services.

Signed-off-by: Logan Seeley <seeley@bitwise.io>