Add orchestrated service endpoints to splinterd
Services that are started using the service orchestrator may need to
expose REST API endpoints for local interactions. To support this, the
orchestrator now creates and supplies endpoints for its services.

Service factories have a new method, `get_rest_endpoints`, that
returns a list of endpoints that the services it creates will need to
have exposed. The parameters for these endpoints are:
- The type of service this endpoint is for
- The route for this endpoint (e.g. "/batches")
- The HTTP method
- A handler function that takes an HTTP request, payload, and a Service

On startup, the orchestrator gets the rest endpoints from all of its
service factories and creates resources from them. Circuit and service
ID parameters are added to the factories' defined routes. When one of
these resources is triggered, the orchestrator will get the speficied
service (using the circuit and service ID parameters) and pass it into
the appropriate factory's handler function along with the request and
payload.

The orchestrator's resources are then added to the splinterd REST API.

Signed-off-by: Logan Seeley <seeley@bitwise.io>