Reorder internal service startup
Reorder the internal service startup to after all the network-level
infrastructure is in place.  This ensures that the admin service and
orchestrator can appropriately connect and startup.

The orchestrator is added to the list of service peer ids and the list
is used to listen for an internal connection for the two expected
services in connection order.

In order to support this change, the network-to-dispatcher loop has to
be pulled into a thread and joined to keep the main thread alive.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>