Add CircuitProposalContext for pending_changes
CircuitProposalContext will be used to show if the pending
changes is a new proposal or a vote for an existing proposal.
This will be used on commit to send the correct AdminServiceEvent.

Also adds a CircuitProposalStatus that will be used as a return
type when checking if a CircuitProposal is ready to be commited.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>