Handle a CircuitProposalVote
The admin service will take the CircuitMangementPayload recieved
from the REST API and validate the proposal vote is valid.

If it is valid the circuit_proposal is updated to have a new
VoteRecord and placed in pending state. It is then passed along
to consensus.

Once consensus agrees to commit the updated circuit_proposal, it will
be committed to open_proposals.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>