Check if CircuitProposal should be committed to SplinterState
When a CircuitProposal is committed, check to see if it has been
approved. Approved means that every member of the proposed circuit,
besides the requester, has submitted an Accept vote.

If accepted the proposed circuit is converted to a SplinterState
Circuit and added to SplinterState. The proposal is then removed.

If the CircuitProposal received a Reject vote, the proposal is removed.

If the CircuitProposal is not yet accepted the CircuitProposal is
replaced in open_proposals with the added VoteRecord.

After the status is determined, an AdminServiceEvent is sent to the
registered Application Authorization Handler, notifying them of the update
and state of the CircuitProposal.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>