Implement workaround for service startup timing
This commit implements a workaround for the timing of service
registration on startup.

2PC commits on the coordinator before instructing the other services to
commit. Because service registration currently requires all nodes to be
available, the coordinator can't startup its scabbard service when 2PC
commits; the other nodes will not have started their services yet (since
they have not yet received the instruction to commit).

The correct solution to this would be to update the node such that on
circuit creation, the node asks other allowed nodes on the circuit if
they have services that are already registered.

This workaround instead adds a "proposal accept callback" to 2PC that is
called when 2PC has finished committing and broadcasting the commit
message to other nodes. The admin service sets this callback to a
closure that will check if the proposal that was just committed was the
last vote needed to commit the circuit; if it is, will initialize the
services for that circuit.

Signed-off-by: Logan Seeley <seeley@bitwise.io>