Remove unused splinter cli actions
This cli started out as a way to test the splinter code
during development (network echo, service commands, etc..).
After the addition of authorization however, the commands no
longer worked because the cli did authorize with the splinter
network and there was no splinter REST API to communicate with.

With the addition of admin key commands, the unused/broken commands
should be removed to make way for a useful splinter cli.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>