Remove inaccurate ursa-compat warn log message
The admin service now takes a signature verifier in as
an argument so no longer needs to check if the ursa-compat
features is enabled.

We are also currently using a sawtooth signer instead of ursa
due to dependency issue with the version of secp256k1
used by ursa and the sawtooth-sdk.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>