Make transport::rw::read/write blocking
The higher-level code, such as mesh, should not need to concern itself
about low-level I/O details such as assembling the bytes of a message.

This commit is a hack and adds in a couple of serious issues because of
the sleep approach. This is a short-term fix to the existing code, which
will have to be refactored to rely less directly on mio to resolve
because mio polling is too low-level (essentially, we need polling at
the message level, not file descriptor level). As a result, performance
after this commit is extremely poor. However, it is more functionally
correct.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>