Add circuit id to ServiceFacotory.create
Although the service implementation is not required to know
about what circuit it is running on, the factory should provide
this information in case the service requires this information.

For example to create a unique name for its database made out of
its service id and the circuit it is running on. There is no
guarantee that a service id is unique outside of the circuit.
To create a unique but reproducable database name, the circuit
id is required.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>