Replace uuid with service_id::circuit_id in Scabbard db name
On restart of a scabbard service, it would genearte a new uuid
and thus create a new database instead of connecting to the
one that already exists.

The name is switched to be service_id::circuit_id because this
combination is guaranteed to be unique on a splinter node but
reproducable on splinter daemon restart.

A service id is not sufficent becasue this is only required to
be unique with in the a circuit, and another service could exist
with the same id in a different circuit.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>