Use ConfigBuilder for loading Toml config files
The approach taken here is to create a TomlConfig instance and use it to
populate ConfigBuilder. TomlConfig is introduced as a separate struct
primarily because it is directly used when loading the toml file and
thus must match its structure. Thus TomlConfig, more than Config, must
be concerned with backward compatibility.

This is gated behind the experimental features "config-toml".

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>