Add BatchHistory to scabbard state
Adds an object to scabbard state that keeps track of submitted batches.
BatchHistory keeps track of the 100 most recently submitted batches.
Batches start off as pending until they are scheduled to be
committed, at which point they become either valid or invalid. Valid
batches are changed to committed when ScabbardState::commit is called.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>