Use game name hash in games route
There is a bug that causes the vue router to break if `%` is present in 
the game name when accessing the route 
`/gamerooms/<circuit_id>/games/<url_encoded_game_name>`. To go around 
this bug instead of using the url encoded game name in the route, a md5 
hash of the game name will be used. 

Signed-off-by: Eloá Franca Verona <eloafran@bitwise.io>