Remove run_interval from EventDealerWebSocket
This commit removes the run_interval usage and replaces it with a
combination of futures library channels and a StreamHandler
implementation.

This uses the underlying futures Stream trait (as implemented by the
UnboundedReceiver) to directly pipe the values to the websocket actor.

This improves the perceived time of operations as the previous
implementation had combined issues of using long interval and only
operating on a single message at a time.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>