Remove Service Connect/Disconnect Forward messages
These messages where used to tell other members of a
circuit where a service was connected before the
list of allowed nodes was added to the circuit defination.

Catchup or discovery was missing from this mechanism and
nodes/services could get out of sink about where services
were located.

The other handlers will be updated to assume the service is
where it is allowed to be in the circuit defination.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>