Update connect on restart logic to check for local node id
During restart, every node stored in splinter state is
reconnected to. That node list will include the local node
and the node should not try to connect to itself. Before
this check was done by checking the endpoints. This should
instead check if the node's node id is the same as the local
node id and skip the connection attempt.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>