Only exit the transport Listener thread on IoError
Only exit if there was an IoError via listening for incoming
connections.  This error should also be logged, for future debugging
purposes (in order to fine-tune the response to these errors).

Likewise, the loop should not be exited if the connection cannot be
added to network (though, this probably indicates that other parts of
the system - mesh, specifically - are shutting down).

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>