Log failures of REST API start-up errors
Log any errors that may occur during the REST API startup, which is
processed in a background thread.  Any errors that were previously
returned as StartUpError will be logged and the thread exited.

This will prevent silent issues from occurring that are not detected
until the REST API is terminated.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>