Update scabbard StateChangeEvent format
Update the state change event emitted by scabbard to be a collection of
individual state changes, correlated by an ID. Each of these events will
correspond to a TransactionReceipt that is stored by the scabbard
service, and the ID of the event will be the ID of the corresponding
receipt.

Also update the scabbard event type to be a single StateChangeEvent
rather than a vector of StateChangeEvents to simplify and clarify the
event dealer.

Signed-off-by: Logan Seeley <seeley@bitwise.io>