Add traits for durable sets
Durable Sets are collections of data backed by a durable store.  These
can come in two flavors - unordered or ordered. The ordered version is
indexed to a particular value which should be derived from the primary
storage type.

This commit provides an pair of in-memory implementations, backed by
HashSet (for the unordered version) and a BTreeSet (for the ordered
version). These versions supply tests that can be treated as reference
implementations.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>