Add admin event mailbox
This commit introduces the idea of an admin event mailbox.  The mail box
will store all of the admin service events in a ordered durable set,
where the natural order is based on the timestamp of the event.

All events may be stored in the mailbox, regardless of circuit mangement type.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>