Decouple EventDealer and EventHistory
Decouple the event history from the event dealer.  This will allow the
storage of events to be managed seperately from the event history, where
items can be stored in a persistence-friendly format.  Likewise, the
stored values do not have to match the serialization format for the
events themselves, but can use whatever format new underlying storage
requires.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>