Add POST /nodes route to splinterd
Add the POST /nodes route to the splinterd REST API for adding nodes to
the local node registry. The endpoint takes the node serialized as JSON.
If a node with the given ID already exists, a 405 Forbidden error is
returned; if the node is invalidly formatted, a 400 Bad Request error is
returned.

Signed-off-by: Logan Seeley <seeley@bitwise.io>