Add cert generate subcommand to splinter cli
This subcommand will generate certificates that
can be used for development. They should never be used
in production.

There are options to change where the certs and keys will
be generated. If this is not provided, an environment
variable, SPLINTER_CERT_DIR, will be checked. Other
wise the command will try to write the certs/keys to
/etc/splinter/certs/.

If the files already exist the command will
fail unless --force is provided. You can also use
--skip to see if the files exist and generate them
if they do not.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>